hw3/flight-service/requirements.txt
grpcio==1.59.2
grpcio-tools==1.59.2
protobuf==4.25.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
tenacity==8.2.3